first working version refactored the schema to its own file
